% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-package.R
\docType{package}
\name{splcr-package}
\alias{splcr}
\alias{splcr-package}
\title{splcr: Smoothed Predictive Likelihood for Competing Risks}
\description{
Internal package metadata and native routine registration.
}
\author{
\strong{Maintainer}: Hao Chen \email{hao.chen@sydney.edu.au}

Authors:
\itemize{
  \item Clara Grazian
}

}
\keyword{internal}
